Flyttet ut tester knyttet til henlegg til behandlingControllerTest, rettet opp i feil knyttet til lesbarhet
