+++ src/test/kotlin/ApplicationLocalPostgres.kt
@@ -26,8 +26,8 @@ fun main(args: Array<String>) {
             "mock-inntekt",
             "mock-ereg",
             "mock-aareg",
-            "mock-brev",
-            "mock-dokument",
+            //"mock-brev",
+            //"mock-dokument",
             "mock-tilbakekreving",
             "mock-klage",
             "mock-sigrun",
