+++ src/test/kotlin/no/nav/familie/ef/sak/behandling/henlegg/HenleggBehandlingControllerTest.kt
@@ -3,8 +3,14 @@ package no.nav.familie.ef.sak.behandling.henlegg
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
+import no.nav.familie.ef.sak.OppslagSpringRunnerTest
+import no.nav.familie.ef.sak.behandling.BehandlingRepository
+import no.nav.familie.ef.sak.behandling.domain.BehandlingResultat
+import no.nav.familie.ef.sak.behandling.domain.BehandlingType
+import no.nav.familie.ef.sak.behandling.dto.BehandlingDto
 import no.nav.familie.ef.sak.behandling.dto.HenlagtDto
 import no.nav.familie.ef.sak.behandling.dto.HenlagtÅrsak
+import no.nav.familie.ef.sak.fagsak.domain.PersonIdent
 import no.nav.familie.ef.sak.felles.util.BrukerContextUtil.clearBrukerContext
 import no.nav.familie.ef.sak.felles.util.BrukerContextUtil.mockBrukerContext
 import no.nav.familie.ef.sak.infrastruktur.exception.Feil
@@ -21,17 +27,28 @@ import no.nav.familie.ef.sak.opplysninger.personopplysninger.dto.Personopplysnin
 import no.nav.familie.ef.sak.opplysninger.personopplysninger.dto.SivilstandDto
 import no.nav.familie.ef.sak.opplysninger.personopplysninger.dto.UtflyttingDto
 import no.nav.familie.ef.sak.opplysninger.personopplysninger.dto.VergemålDto
+import no.nav.familie.ef.sak.repository.behandling
+import no.nav.familie.ef.sak.repository.fagsak
 import no.nav.familie.ef.sak.vilkår.dto.StatsborgerskapDto
+import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.familie.prosessering.internal.TaskService
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.web.client.exchange
+import org.springframework.http.HttpEntity
+import org.springframework.http.HttpMethod
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import java.time.LocalDate
 import java.util.UUID
 
-class HenleggBehandlingControllerTest {
+class HenleggBehandlingControllerTest : OppslagSpringRunnerTest() {
+    @Autowired
+    private lateinit var behandlingRepository: BehandlingRepository
     val taskService = mockk<TaskService>(relaxed = true)
     val personopplysningerService = mockk<PersonopplysningerService>(relaxed = true)
 
@@ -48,6 +65,7 @@ class HenleggBehandlingControllerTest {
 
     @BeforeEach fun setUp() {
         mockBrukerContext("", azp_name = "prod-gcp:teamfamilie:familie-ef-personhendelse")
+        headers.setBearerAuth(lokalTestToken)
     }
 
     @AfterEach
@@ -83,7 +101,7 @@ class HenleggBehandlingControllerTest {
     }
 
     @Test internal fun `Skal ikke kaste feil hvis bruker har fullmakt som har utgått`() {
-        mocckHentPersonopplysningerMedFullmaktEttÅrSiden()
+        mocckHentPersonopplysningerMedFullmaktEnDagSiden()
         henleggBehandlingController.henleggBehandling(UUID.randomUUID(), HenlagtDto(HenlagtÅrsak.TRUKKET_TILBAKE, true))
         verify(exactly = 1) { taskService.save(any()) }
     }
@@ -97,7 +115,46 @@ class HenleggBehandlingControllerTest {
         assertThat(exception.message).isEqualTo("Skal ikke sende brev hvis person er tilknyttet vergemål eller fullmakt")
     }
 
-    private fun mocckHentPersonopplysningerMedFullmaktEttÅrSiden() {
+    @Test
+    internal fun `Skal henlegge behandling`() {
+        val fagsak = testoppsettService.lagreFagsak(fagsak(identer = setOf(PersonIdent("12345678901"))))
+        val behandling = behandlingRepository.insert(behandling(fagsak, type = BehandlingType.FØRSTEGANGSBEHANDLING))
+        val respons = henlegg(behandling.id, HenlagtDto(årsak = HenlagtÅrsak.FEILREGISTRERT))
+
+        assertThat(respons.statusCode).isEqualTo(HttpStatus.OK)
+        assertThat(respons.body?.data!!.resultat).isEqualTo(BehandlingResultat.HENLAGT)
+        assertThat(respons.body?.data!!.henlagtÅrsak).isEqualTo(HenlagtÅrsak.FEILREGISTRERT)
+    }
+
+    @Test
+    internal fun `Skal henlegge FØRSTEGANGSBEHANDLING`() {
+        val fagsak = testoppsettService.lagreFagsak(fagsak(identer = setOf(PersonIdent("12345678901"))))
+        val behandling = behandlingRepository.insert(behandling(fagsak, type = BehandlingType.FØRSTEGANGSBEHANDLING))
+        val respons = henlegg(behandling.id, HenlagtDto(årsak = HenlagtÅrsak.FEILREGISTRERT))
+
+        assertThat(respons.statusCode).isEqualTo(HttpStatus.OK)
+        assertThat(respons.body?.data!!.resultat).isEqualTo(BehandlingResultat.HENLAGT)
+        assertThat(respons.body?.data!!.henlagtÅrsak).isEqualTo(HenlagtÅrsak.FEILREGISTRERT)
+    }
+
+    private fun henleggBehandling(id: UUID): ResponseEntity<Ressurs<BehandlingDto>> =
+        restTemplate.exchange(
+            localhost("/api/behandling/$id/henlegg"),
+            HttpMethod.POST,
+            HttpEntity<Ressurs<BehandlingDto>>(headers),
+        )
+
+    private fun henlegg(
+        id: UUID,
+        henlagt: HenlagtDto,
+    ): ResponseEntity<Ressurs<BehandlingDto>> =
+        restTemplate.exchange<Ressurs<BehandlingDto>>(
+            localhost("/api/behandling/$id/henlegg"),
+            HttpMethod.POST,
+            HttpEntity(henlagt, headers),
+        )
+
+    private fun mocckHentPersonopplysningerMedFullmaktEnDagSiden() {
         every { personopplysningerService.hentPersonopplysningerFraRegister(any()) } returns
             dto(
                 fullmakt =
@@ -178,3 +235,4 @@ class HenleggBehandlingControllerTest {
         vergemål = vergemål,
     )
 }
+
